[WV-2287] Moved most logic back to Import and Invite page [TEAM REVIEW]
